feat(dashboard): created Dashboard page with TempLayout, added 2 charts packages
