+++ src/mocks/lastOrdersMock.ts
@@ -1,0 +1,148 @@
+export default {
+  headers: [
+    { title: 'Customer', key: 'customer' },
+    { title: 'Order', key: 'order' },
+    { title: 'Delivery Date', key: 'deliveryDate' },
+    { title: 'Delivery Pricing', key: 'deliveryPricing' },
+    { title: 'Delivery Status', key: 'deliveryStatus' },
+    { title: 'Payment', key: 'payment' }
+  ],
+  orders: [
+    {
+      customer: 'John Doe',
+      order: 'Pantalones',
+      deliveryDate: '2023-07-21',
+      deliveryPricing: '$12.99',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'Jane Smith',
+      order: 'Camisas',
+      deliveryDate: '2023-07-22',
+      deliveryPricing: 'Free',
+      deliveryStatus: 'Pending',
+      payment: 'Transfer'
+    },
+    {
+      customer: 'Mike Johnson',
+      order: 'Gorras',
+      deliveryDate: '2023-07-23',
+      deliveryPricing: '$8.50',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'Alice Williams',
+      order: 'Vestidos',
+      deliveryDate: '2023-07-24',
+      deliveryPricing: '$9.99',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'Bob Anderson',
+      order: 'Zapatos',
+      deliveryDate: '2023-07-25',
+      deliveryPricing: '$15.00',
+      deliveryStatus: 'Pending',
+      payment: 'Transfer'
+    },
+    {
+      customer: 'Eva Martinez',
+      order: 'Bolsos',
+      deliveryDate: '2023-07-26',
+      deliveryPricing: '$7.50',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'David Lee',
+      order: 'Sombreros',
+      deliveryDate: '2023-07-27',
+      deliveryPricing: '$6.99',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'Sophia Kim',
+      order: 'Sweaters',
+      deliveryDate: '2023-07-28',
+      deliveryPricing: '$11.50',
+      deliveryStatus: 'Pending',
+      payment: 'Transfer'
+    },
+    {
+      customer: 'Michael Brown',
+      order: 'Chaqueta',
+      deliveryDate: '2023-07-29',
+      deliveryPricing: '$14.75',
+      deliveryStatus: 'Canceled',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'Olivia Davis',
+      order: 'Pantalones cortos',
+      deliveryDate: '2023-07-30',
+      deliveryPricing: '$8.25',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'William Wilson',
+      order: 'Bufandas',
+      deliveryDate: '2023-07-31',
+      deliveryPricing: '$5.00',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'Ava Taylor',
+      order: 'Faldas',
+      deliveryDate: '2023-08-01',
+      deliveryPricing: '$10.00',
+      deliveryStatus: 'Pending',
+      payment: 'Transfer'
+    },
+    {
+      customer: 'James Johnson',
+      order: 'Corbatas',
+      deliveryDate: '2023-08-02',
+      deliveryPricing: '$6.50',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'Emma Martin',
+      order: 'Calcetines',
+      deliveryDate: '2023-08-03',
+      deliveryPricing: '$3.99',
+      deliveryStatus: 'Canceled',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'Alexander Martinez',
+      order: 'Gafas',
+      deliveryDate: '2023-08-04',
+      deliveryPricing: '$9.50',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    },
+    {
+      customer: 'Charlotte Thompson',
+      order: 'Camisetas',
+      deliveryDate: '2023-08-05',
+      deliveryPricing: 'Free',
+      deliveryStatus: 'Pending',
+      payment: 'Transfer'
+    },
+    {
+      customer: 'Daniel Hall',
+      order: 'Sandalias',
+      deliveryDate: '2023-08-06',
+      deliveryPricing: '$12.00',
+      deliveryStatus: 'Complete',
+      payment: 'Credit Card'
+    }
+  ]
+}
