+++ vite.config.ts
@@ -5,9 +5,6 @@ import vue from '@vitejs/plugin-vue'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  server: {
-    open: true
-  },
   plugins: [vue()],
   resolve: {
     alias: {
